Fix type consistency across multiple modules
- configTimeToCycles: return uint16_t instead of int32_t
- dataPackPacked: return uint8_t instead of int8_t
- uartPutsNonBlocking: change len from uint16_t to size_t
- Fix loop counter types (int32_t -> uint8_t/uint32_t)

+++ src/configuration.c
@@ -51,7 +51,7 @@ static void     configDefault(void);
 static void     configEchoQueueChar(const uint8_t c);
 static void     configEchoQueueStr(const char *s);
 static void     configInitialiseNVM(void);
-static int32_t  configTimeToCycles(const float time, const int32_t mainsFreq);
+static uint16_t configTimeToCycles(const float time, const int32_t mainsFreq);
 static bool     configureAnalog(void);
 static bool     configureAssumed(void);
 static void     configureBackup(void);
@@ -1469,8 +1469,8 @@ void configProcessCmd(void) {
 
 bool configUnsavedChanges(void) { return unsavedChange; }
 
-int32_t configTimeToCycles(const float time, const int32_t mainsFreq) {
-  return qfp_float2uint(qfp_fmul(time, qfp_int2float(mainsFreq)));
+uint16_t configTimeToCycles(const float time, const int32_t mainsFreq) {
+  return (uint16_t)qfp_float2uint(qfp_fmul(time, qfp_int2float(mainsFreq)));
 }
 
 VersionInfo_t configVersion(void) {
